[fix][adj] issue-1. Dataset_dir can not found; issue-2. ram_size error fix:     src/booking/booking.py, issue-1, fixed     src/monitor/run_container.py, issue-2, fixed adj:     src/HostInfo.py, shm_rate: float

+++ src/HostInfo.py
@@ -97,7 +97,7 @@ class MaxCapability:
     swap_size: int
     gpus: int
 
-    shm_rate: int  # swap_size / ram if swap_size != 0 else 1
+    shm_rate: float  # ram / (ram + swap_size)
     memory: int  # ram + swap_size
 
     def __init__(self, max_dict: dict) -> None:
